spelling, cleanup, check for null chars in packages
