+++ src/AppInstallerRepositoryCore/AppInstallerRepositoryCore.vcxproj
@@ -391,7 +391,6 @@
     <ClInclude Include="Microsoft\Schema\2_0\PackageFamilyNameTable.h" />
     <ClInclude Include="Microsoft\Schema\2_0\PackagesTable.h" />
     <ClInclude Include="Microsoft\Schema\2_0\OneToManyTableWithMap.h" />
-    <ClInclude Include="Microsoft\Schema\2_0\OneToOneTable.h" />
     <ClInclude Include="Microsoft\Schema\2_0\ProductCodeTable.h" />
     <ClInclude Include="Microsoft\Schema\2_0\SearchResultsTable.h" />
     <ClInclude Include="Microsoft\Schema\2_0\SystemReferenceStringTable.h" />
@@ -498,8 +497,7 @@
     <ClCompile Include="Microsoft\Schema\1_7\Interface_1_7.cpp" />
     <ClCompile Include="Microsoft\Schema\2_0\Interface_2_0.cpp" />
     <ClCompile Include="Microsoft\Schema\2_0\PackagesTable.cpp" />
-    <ClCompile Include="Microsoft\Schema\2_0\OneToManyTableWithMap_2_0.cpp" />
-    <ClCompile Include="Microsoft\Schema\2_0\OneToOneTable_2_0.cpp" />
+    <ClCompile Include="Microsoft\Schema\2_0\OneToManyTableWithMap.cpp" />
     <ClCompile Include="Microsoft\Schema\2_0\SearchResultsTable_2_0.cpp" />
     <ClCompile Include="Microsoft\Schema\2_0\SystemReferenceStringTable.cpp" />
     <ClCompile Include="Microsoft\Schema\ISQLiteIndex.cpp" />
