+++ src/AppInstallerRepositoryCore/Microsoft/Schema/1_6/Interface_1_6.cpp
@@ -98,7 +98,7 @@ namespace AppInstaller::Repository::Microsoft::Schema::V1_6
 
     void Interface::DropTables(SQLite::Connection& connection)
     {
-        SQLite::Savepoint savepoint = SQLite::Savepoint::Create(connection, "droptables_v1_6");
+        SQLite::Savepoint savepoint = SQLite::Savepoint::Create(connection, "drop_tables_v1_6");
 
         V1_4::Interface::DropTables(connection);
